+++ Competitive_Programming/Computation_Basic/mod_combination.hpp
@@ -3,8 +3,7 @@
 #define MAX_N 100005
 #define MOD 1000000007
 
-//どこかでmake()することを忘れずに
-ll inv[MAX_N],fac[MAX_N],finv[MAX_N];
+int inv[MAX_N],fac[MAX_N],finv[MAX_N];
 
 void make()
 {
@@ -12,16 +11,16 @@ void make()
 	finv[0] = finv[1] = 1;
 	inv[1] = 1;
 	for(int i=2;i<MAX_N;i++){
-		inv[i] = MOD - inv[MOD%i] * (MOD/i) % MOD;
-		fac[i] = fac[i-1] * (ll) i % MOD;
-		finv[i] = finv[i-1] * inv[i] % MOD;
+		inv[i] = MOD - (long long)inv[MOD%i] * (MOD/i) % MOD;
+		fac[i] = (long long)fac[i-1] * i % MOD;
+		finv[i] = (long long)finv[i-1] * inv[i] % MOD;
 	}
 }
 
-ll comb(int a,int b)
+int comb(int a,int b)
 {
-	if(a<b){
+	if(a < b){
 		return 0;
 	}
-	return fac[a] * (finv[b] * finv[a-b] % MOD) % MOD;
+	return fac[a] * ((long long)finv[b] * finv[a-b] % MOD) % MOD;
 }
