+++ Competitive_Programming/Computation_Basic/mod_pow.hpp
@@ -2,10 +2,10 @@
 
 #define MOD 1000000007
 
-ll mod_pow(ll a,ll b)
+long long mod_pow(long long a,long long b)
 {
     a %= MOD;
-    ll res = 1;
+    long long res = 1;
     while(b){
         if(b & 1){
             res = res * a % MOD;
