+++ Competitive_Programming/Computation_Basic/prime_factorize.hpp
@@ -1,10 +1,10 @@
 #include "../header.hpp"
 
-vector<ll> dev;
+vector<long long> dev;
 vector<int> id;
 
-void prime_factor(ll N){
-	for(ll i=2;i*i<=N;i++){
+void prime_factor(long long N){
+	for(int i=2;(long long)i*i<=N;i++){
 		int cnt = 0;
 		while(N%i == 0){
 			cnt++;
