+++ Competitive_Programming/Computation_Basic/prime_judge.hpp
@@ -1,9 +1,9 @@
 #include "../header.hpp"
 
 //素数かどうかを判定する
-bool is_prime(ll n)
+bool is_prime(long long n)
 {
-    for(ll i=2;i*i<=n;i++){
+    for(int i=2;(long long)i*i<=n;i++){
         if(n % i == 0){
             return false;
         }
